+++ docs/changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The script `reprep-sync-zenodo` and corresponding StepUp API function `sync_zenodo()`
   synchronize your local data with a draft dataset on Zenodo.
+- Small documentation updates
 
 
 ## [1.2.1] - 2024-05-27