+++ docs/changelog.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   synchronize your local data with a draft dataset on Zenodo.
 - Small documentation updates
 
+### Fixed
+
+- Upgraded dependency markdown-katex to version 202406.1035,
+  and enabled concurrency for markdown conversion with equations.
+
 
 ## [1.2.1] - 2024-05-27 {: #v1.2.1 }